refactor Java libs: make Docker the default executor
