+++ localstack/ext/java/src/main/java/cloud/localstack/LocalstackTestRunner.java
@@ -1,19 +1,34 @@
 package cloud.localstack;
 
+import cloud.localstack.docker.annotation.LocalstackDockerAnnotationProcessor;
+import cloud.localstack.docker.annotation.LocalstackDockerConfiguration;
 import org.junit.runner.notification.RunNotifier;
 import org.junit.runners.BlockJUnit4ClassRunner;
 import org.junit.runners.model.InitializationError;
 
 /**
- * Simple JUnit test runner that automatically downloads, installs, starts,
- * and stops the LocalStack local cloud infrastructure components.
+ * JUnit test runner that automatically pulls and runs the latest localstack docker image
+ * and then terminates when tests are complete.
  *
- * Should work cross-OS, however has been only tested under Unix (Linux/MacOS).
+ * Having docker installed is a prerequisite for this test runner to execute.  If docker
+ * is not installed in one of the default locations (C:\program files\docker\docker\resources\bin\, usr/local/bin or
+ * usr/bin)
+ * then use the DOCKER_LOCATION environment variable to specify the location.
  *
- * @author Waldemar Hummer
+ * Since ports are dynamically allocated, the external port needs to be resolved based on the default localstack port.
+ *
+ * The hostname defaults to localhost, but in some environments that is not sufficient, so the HostName can be specified
+ * by using the LocalstackDockerProperties annotation with an IHostNameResolver.
+ *
+ * @author Alan Bevier
  * @author Patrick Allain
+ * @author Waldemar Hummer
  */
 public class LocalstackTestRunner extends BlockJUnit4ClassRunner {
+
+    private static final LocalstackDockerAnnotationProcessor PROCESSOR = new LocalstackDockerAnnotationProcessor();
+
+    private Localstack localstackDocker = Localstack.INSTANCE;
 
     public LocalstackTestRunner(Class<?> klass) throws InitializationError {
         super(klass);
@@ -21,8 +36,13 @@ public class LocalstackTestRunner extends BlockJUnit4ClassRunner {
 
     @Override
     public void run(RunNotifier notifier) {
-        Localstack.INSTANCE.setupInfrastructure();
-        super.run(notifier);
+        try {
+            final LocalstackDockerConfiguration dockerConfig = PROCESSOR.process(this.getTestClass().getJavaClass());
+            localstackDocker.startup(dockerConfig);
+            super.run(notifier);
+        } finally {
+            localstackDocker.stop();
+        }
     }
 
 }
