+++ localstack/ext/java/src/main/java/cloud/localstack/deprecated/TestUtils.java
@@ -1,4 +1,4 @@
-package cloud.localstack;
+package cloud.localstack.deprecated;
 
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSCredentialsProvider;
@@ -6,10 +6,6 @@ import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.client.builder.AwsClientBuilder;
 import com.amazonaws.client.builder.ExecutorFactory;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDB;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBClientBuilder;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBStreams;
-import com.amazonaws.services.dynamodbv2.AmazonDynamoDBStreamsClientBuilder;
 import com.amazonaws.services.kinesis.AmazonKinesis;
 import com.amazonaws.services.kinesis.AmazonKinesisAsync;
 import com.amazonaws.services.kinesis.AmazonKinesisAsyncClientBuilder;
@@ -41,6 +37,8 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Stream;
+
+import static cloud.localstack.TestUtils.getCredentialsProvider;
 
 @SuppressWarnings("all")
 public class TestUtils {
@@ -125,22 +123,18 @@ public class TestUtils {
         return builder.build();
     }
 
+    public static AmazonS3 getClientS3SSL() {
+        AmazonS3ClientBuilder builder = AmazonS3ClientBuilder.standard().
+                withEndpointConfiguration(getEndpointConfigurationS3SSL()).
+                withCredentials(getCredentialsProvider());
+        builder.setPathStyleAccessEnabled(true);
+        return builder.build();
+    }
+
     public static AWSSecretsManager getClientSecretsManager() {
         return AWSSecretsManagerClientBuilder.standard().
                 withEndpointConfiguration(getEndpointConfigurationSecretsManager()).
                 withCredentials(getCredentialsProvider()).build();
-    }
-
-    public static AmazonDynamoDB getClientDynamoDB() {
-        return AmazonDynamoDBClientBuilder.standard()
-                .withEndpointConfiguration(getEndpointConfigurationDynamoDB())
-                .withCredentials(getCredentialsProvider()).build();
-    }
-
-    public static AmazonDynamoDBStreams getClientDynamoDBStreams() {
-        return AmazonDynamoDBStreamsClientBuilder.standard()
-                .withEndpointConfiguration(getEndpointConfigurationDynamoDBStreams())
-                .withCredentials(getCredentialsProvider()).build();
     }
 
     public static AmazonKinesis getClientKinesis() {
@@ -163,39 +157,35 @@ public class TestUtils {
     }
 
     protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationLambda() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointLambda());
+        return getEndpointConfiguration(Localstack.getEndpointLambda());
     }
 
     protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationKinesis() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointKinesis());
-    }
-
-    protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationDynamoDB() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointDynamoDB());
-    }
-
-    protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationDynamoDBStreams() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointDynamoDBStreams());
+        return getEndpointConfiguration(Localstack.getEndpointKinesis());
     }
 
     protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationSQS() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointSQS());
+        return getEndpointConfiguration(Localstack.getEndpointSQS());
     }
 
     protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationS3() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointS3());
+        return getEndpointConfiguration(Localstack.getEndpointS3());
     }
 
     protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationSNS() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointSNS());
+        return getEndpointConfiguration(Localstack.getEndpointSNS());
+    }
+
+    protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationS3SSL() {
+        return getEndpointConfiguration(Localstack.getEndpointS3(true));
     }
 
     protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationSecretsManager() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointSecretsmanager());
+        return getEndpointConfiguration(Localstack.getEndpointSecretsmanager());
     }
 
     protected static AwsClientBuilder.EndpointConfiguration getEndpointConfigurationStepFunctions() {
-        return getEndpointConfiguration(Localstack.INSTANCE.getEndpointStepFunctions());
+        return getEndpointConfiguration(Localstack.getEndpointStepFunctions());
     }
 
     protected static void setEnv(Map<String, String> newEnv) {
