+++ localstack/ext/java/src/test/java/cloud/localstack/S3FeaturesTest.java
@@ -73,9 +73,10 @@ public class S3FeaturesTest {
 
 		String bucketName = "test-bucket-https";
 
+		System.out.println(Localstack.INSTANCE.getEndpointS3());
 		AmazonS3 amazonS3Client = AmazonS3ClientBuilder.standard()
 				.withEndpointConfiguration(new AwsClientBuilder.EndpointConfiguration(
-						Localstack.getEndpointS3(),
+						Localstack.INSTANCE.getEndpointS3(),
 						Localstack.getDefaultRegion()))
 				.withCredentials(TestUtils.getCredentialsProvider())
 				.withChunkedEncodingDisabled(true)
