+++ localstack/ext/java/src/test/java/cloud/localstack/deprecated/BasicFunctionalityTest.java
@@ -1,9 +1,10 @@
-package cloud.localstack;
+package cloud.localstack.deprecated;
 
+import cloud.localstack.LocalTestUtil;
 import cloud.localstack.sample.KinesisLambdaHandler;
 import cloud.localstack.sample.S3Sample;
 import cloud.localstack.sample.SQSLambdaHandler;
-import cloud.localstack.sample.SQSLambdaHandlerSSL;
+
 import com.amazonaws.ClientConfiguration;
 import com.amazonaws.Protocol;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
@@ -54,7 +55,7 @@ import static cloud.localstack.TestUtils.TEST_CREDENTIALS;
  *
  * @author Waldemar Hummer
  */
-@RunWith(LocalstackTestRunner.class)
+@RunWith(LocalstackOutsideDockerTestRunner.class)
 @ExtendWith(LocalstackExtension.class)
 public class BasicFunctionalityTest {
 
@@ -146,13 +147,8 @@ public class BasicFunctionalityTest {
         request.setFunctionName(functionName);
         request.setRuntime(Runtime.Java8);
         request.setRole("r1");
-        if(Localstack.useSSL()){
-            request.setCode(LocalTestUtil.createFunctionCode(SQSLambdaHandlerSSL.class));
-            request.setHandler(SQSLambdaHandlerSSL.class.getName());
-        } else {
-            request.setCode(LocalTestUtil.createFunctionCode(SQSLambdaHandler.class));
-            request.setHandler(SQSLambdaHandler.class.getName());
-        }
+        request.setCode(LocalTestUtil.createFunctionCode(SQSLambdaHandler.class));
+        request.setHandler(SQSLambdaHandler.class.getName());
         lambda.createFunction(request);
 
         // create stream
